Avoid warning about unused function, fixes encukou#40
By marking the static function "inline", the compiler will only generate code for it if it is actually used. Without the keyword, the compiler will emit a warning if it is not used. For trivial Python C bindings this is not an issue, but if the code consists of multiple compilation units then this currently results in a warning for each such compilation unit without module initialization function.